display the decimal point

+++ HomePage/cpu_monitor.cpp
@@ -40,7 +40,7 @@ CPUMonitor::CPUMonitor(QWidget *parent)
     progress->setTitle("CPU");
 }
 
-void CPUMonitor::setPercentValue(const int &value)
+void CPUMonitor::setPercentValue(const float &value)
 {
     progress->setValue(value);
