+++ home_page.h
@@ -65,9 +65,9 @@ private:
 
 private slots:
     void init();
-    void updateCpuPercent(int cpuPercent);
-    void updateMemory(QString memory, int percent);
-    void updateDisk(QString disk, int percent);
+    void updateCpuPercent(float cpuPercent);
+    void updateMemory(QString memory, float percent);
+    void updateDisk(QString disk, float percent);
     void updateNetworkSpeed(QString upload, QString download);
 };
